pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "celebrity-image-finder"
version = "0.1.0"
description = "CLI tool to find and download high-quality face images of celebrities"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "Pillow>=10.0",
    "mediapipe>=0.10",
    "imagehash>=4.3",
    "lyricsgenius>=3.0",
    "spotipy>=2.23",
    "ddgs>=7.0",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
finder = "finder.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
